modified blog controller to authorize user

+++ controllers/blog.controller.js
@@ -1,15 +1,16 @@
 const blogs=require("../database/connection");
 
+//req.body is an object that contains key-value pairs of data submitted in the request body.
 const fetchBlogs=async(req,res)=>{
    try{
         const datas=await blogs.findAll();
-        res.status(200).json({
+         return res.status(200).json({
             message:"all blogs fetched",
             datas:datas,
         })
    }
    catch(error){
-     res.status(500).json({
+      return res.status(500).json({
         message:"failed",
      })
    }
@@ -19,7 +20,7 @@ const singleFetchBlogs= async(req,res)=>{
    try{
        const {id}=req.params;
         const datas=await blogs.findByPk(id);
-        res.status(200).json({
+         return res.status(200).json({
             message:"all blogs fetched",
             datas:datas,
         })
@@ -33,6 +34,8 @@ const singleFetchBlogs= async(req,res)=>{
 
 const addBlogs=async(req,res)=>{
     try{
+      const usersId = req.user.id;
+
       const {blogTitle,blogAuthor,blogContent,blogStatus,blogPublishDate}=req.body;
        
       await blogs.create({
@@ -41,10 +44,14 @@ const addBlogs=async(req,res)=>{
         blogContent,
         blogStatus,
         blogPublishDate,
+        id:usersId,
       })
+      return res.status(201).json({
+        message:"blog added successfully",
+      });
     }
      catch(error){
-       res.status(500).json({
+       return res.status(500).json({
          message:"adding failed",
        })
     }
@@ -52,9 +59,25 @@ const addBlogs=async(req,res)=>{
 
 const editBlogs= async(req,res)=>{
     try{
+        const usersId = req.user.id;
         const {id}=req.params;
         const {blogTitle,blogAuthor,blogContent,blogStatus,blogPublishDate}=req.body;
         
+        //does the blog exist?
+        const blog= await blogs.findByPk(id);
+        if(!blog){
+            return res.status(404).json({
+                message:"blog not found",
+            });
+        };
+
+        //if it exists, whose blog is it?
+        if(blog.id!==usersId && req.user.role!=="admin"){
+            return res.status(403).json({
+                message:"forbidden, you cant edit this blog",
+            });
+        }
+
         await blogs.update({
                 blogTitle,
                 blogAuthor,
@@ -64,7 +87,10 @@ const editBlogs= async(req,res)=>{
               }
                 , {
                     where: {id}
-                });     
+                });  
+                return res.status(200).json({
+                    message:"blog updated successfully",
+                });   
     } catch(error){
         res.status(500).json({
             message:"update failed",
@@ -74,14 +100,30 @@ const editBlogs= async(req,res)=>{
 
 const deleteBlogs= async(req,res)=>{
     try{
+        const usersId = req.user.id;
        const {id}= req.params;
+
+       const blog= await blogs.findByPk(id);
+
+       if(!blog){
+        return res.status(404).json({
+            message:"blog not found",
+        })
+       };
+       //but blog exists.. but whose?
+
+       if(blog.id!==usersId && req.user.role!=="admin"){
+        return res.status(403).json({
+            message:"forbidden, you cant delete this blog",
+        });
+       }
         await blogs.destroy({
             where:{id},
         })
     }catch(error){
         res.status(500).json({
             message:"deletion failed",
-        })
+        });
     }
 }
